refactor(sqlmem): extract a backend interface behind the Manager facade
Phase 2 (postgres schema-per-scope) needs a second storage tier beside the
Phase-1 sqlite-file-per-scope. Restructure the package so the Manager is a
thin facade that validates a statement then delegates the DB work to a
`backend`, with the Phase-1 sqlite logic moved verbatim into a sqliteBackend.

- sqlmem.go: the Manager facade (validate -> delegate), the `backend`
  interface, the shared ScopeKey/QueryResult/ExecResult, and the shared
  collectRows / effectiveQuota / withTimeout helpers.
- sqlite.go: the sqliteBackend — the Phase-1 file-per-scope logic with
  unchanged behavior (LRU handles, path fence, page-size quota).
- validate.go: validateStatement now dispatches by `dialect`; the sqlite
  path is byte-identical (a 2-arg wrapper preserves the test call sites).
- validate_postgres.go: the postgres-dialect escape denies (CREATE EXTENSION
  / ALTER SYSTEM-class DDL and nested server-side functions) — the only gaps
  the shared closed allow-sets do not already cover. Inert until Phase 2
  wires a backend that selects dialectPostgres.

No behavior change for the sqlite tier; all Phase-1 sqlmem tests pass.
